LayoutTests/sputnik/Conformance/12_Statement/12.8_The_break_Statement/S12.8_A6-expected.txt

CONSOLE MESSAGE: line 80: SyntaxError: Label 'LABEL1' is not defined
S12.8_A6

PASS Expected parsing failure

TEST COMPLETE
